isisd: Fix local IP address dependency on link parameter configuration
When the interface IP address is configured first, and then ISIS MPLS-TE is enabled, the link parameters of the interface and the LSP consistently lack the local IP address.
The reason is that when calling isis_link_params_update(), it checks if the link parameters are enabled through HAS_LINK_PARAMS(ifp). Only when it is enabled does it set the sub-TLV for the local IP address. However, the sub-TLV for the local IP address is not actually dependent on the link parameter configuration. For example, when configuring the interface IP by calling isis_circuit_add_addr(), it only checks if ISIS MPLS-TE is enabled, without checking if the link parameters are enabled.


Signed-off-by: zhou-run <zhou.run@h3c.com>